source-code/requirements.txt
asgiref==3.6.0
channels==3.0.5
channels_redis==3.4.1
Django==3.2
djangorestframework==3.14.0
